Run all the tests in parallel
Previously, all tests were run serially. This is super wasteful of the
GitHub (and human) minutes, because even in production, the VMs are
running on machines that are shared with hundreds of other VMs, which
means that if we manage to get CPU affinity set to a single core per
test, it should approximate the runner to a certain degree.

So let's try running tests in parallel, for *EXTREME SPEED*. If we don't
like it, `-j1` should revert back to the old behavior.